Add the Brocadinator: Cisco config goes in, Brocade config comes out.
Can't explain that.

this is a rough first version but it did get us through the conversion of every switch room in VBRH except M-10C which was done manually. I already know that this will blow up if the input config comes from a 3560G but it has been tested to work with 3500XL, 3550, and 3560-X configs in its current iteration.

git-svn-id: svn://dev.uah.edu/main/Misc/trunk/Scripts@4453 a68432d1-e60b-4b54-a684-c3d375a5ebf5