Add script to convert Brocade "show vlan" to a list of port IDs.
You can supply one argument, which is a string to prepend to each
line full of port ids. So if you do 'show vlan 4000' and piped it
into this script, and that VLAN has ports 1/1/1, 1/1/2, and
2/1/37 in it, and you provided the argument "no untag" to this
script, you will get output like:

no untag eth 1/1/1 eth 1/1/2
no untag eth 2/1/37

The line break is because that is the same place the line is
broken by the 'show vlan' command on the Brocade unit.

git-svn-id: svn://dev.uah.edu/main/Misc/trunk/Scripts@4686 a68432d1-e60b-4b54-a684-c3d375a5ebf5