Add Nagios check for checking LAG member ports over SNMP
This check will use SNMP gets to check whether some ports
in a Brocade ICX LAG have gone down. If 0 ports are up,
this is currently an acceptable condition (since,
presumably, if there was an outage, a host check would
catch that condition). If all of the ports are up, that's
fine too. But if a LAG is running with fewer than its
configured number of ports, the check will return
WARNING.